\name{add.emails}
\alias{add.emails}
\title{Add emails to the account}
\usage{
  add.emails(content, ctx = get.github.context())
}
\arguments{
  \item{content}{either an email or a list of emails}

  \item{ctx}{the github context object}
}
\value{
  The new list of emails
}
\description{
  Add emails to the account
}

