\name{am.following.user}
\alias{am.following.user}
\title{Test whether current user is following given user}
\usage{
  am.following.user(user, ctx = get.github.context())
}
\arguments{
  \item{user}{the given user}

  \item{ctx}{the github context object}
}
\value{
  TRUE or FALSE
}
\description{
  Test whether current user is following given user
}

