% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Api.R
\name{Api}
\alias{Api}
\title{R6 Class for GitHub API object}
\description{
API class representing GitHub API object used to maniuplate users and repos.
}
\details{
Contains two fields \strong{token} and \strong{user} field, which
is of type `User` class. This class can later be used to change repo and user
parameters (such as: description, collaborators etc.)
}
\examples{
\dontrun{
api <- User$new(token = "token", username = "username")
api <- User$new("token", "username")
}

}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{token}}{string Token of the GitHub user}

\item{\code{user}}{User class object.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Api$new()}}
\item \href{#method-print}{\code{Api$print()}}
\item \href{#method-list_repositories}{\code{Api$list_repositories()}}
\item \href{#method-clone}{\code{Api$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a new user object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Api$new(token = NA, username = NA)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{token}}{string Token for the GitHub account.}

\item{\code{username}}{string Username of the GitHub account.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new `Api` object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\if{latex}{\out{\hypertarget{method-print}{}}}
\subsection{Method \code{print()}}{
Change object printing
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Api$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-list_repositories"></a>}}
\if{latex}{\out{\hypertarget{method-list_repositories}{}}}
\subsection{Method \code{list_repositories()}}{
Listing all repositories which the authenticated user has access to
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Api$list_repositories()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Vector of `Repository` objects.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Api$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
