\name{conceal.member.of.organization}
\alias{conceal.member.of.organization}
\title{conceal user membership in an organization}
\usage{
  conceal.member.of.organization(org, user,
    ctx = get.github.context())
}
\arguments{
  \item{org}{the organization name}

  \item{user}{the user name}

  \item{ctx}{the github context object}
}
\value{
  None
}
\description{
  conceal user membership in an organization
}

