% Generated by roxygen2 (4.0.2): do not edit by hand
\name{create.key}
\alias{create.key}
\title{Add a public key with some id}
\usage{
create.key(content, ctx = get.github.context())
}
\arguments{
\item{content}{the public key}

\item{ctx}{the github context object}
}
\value{
the added public key
}
\description{
Add a public key with some id
}

