% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repositories.R
\name{create.organization.repository}
\alias{create.organization.repository}
\title{create rpository under given organization}
\usage{
create.organization.repository(org, ..., ctx = get.github.context())
}
\arguments{
\item{org}{the given organization}

\item{...}{parameters, "name=repo_name" is required, see
\url{http://developer.github.com/v3/repos/#create} for more}

\item{ctx}{the github context object}
}
\value{
the created repository
}
\description{
create rpository under given organization
}

