% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repositories.R
\name{create.repository}
\alias{create.repository}
\title{create repository}
\usage{
create.repository(..., ctx = get.github.context())
}
\arguments{
\item{...}{parameters, "name=repo_name" is required, see
\url{http://developer.github.com/v3/repos/#create} for more}

\item{ctx}{the github context object}
}
\value{
the created repository
}
\description{
create repository
}

