% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/users.R
\name{delete.emails}
\alias{delete.emails}
\title{Delete emails from the account}
\usage{
delete.emails(content, ctx = get.github.context())
}
\arguments{
\item{content}{either an email or a list of emails}

\item{ctx}{the github context object}
}
\value{
Nothing
}
\description{
Delete emails from the account
}

