% Generated by roxygen2 (4.0.2): do not edit by hand
\name{delete.gist.comment}
\alias{delete.gist.comment}
\title{Delete gist comment}
\usage{
delete.gist.comment(gist.id, comment.id, ctx = get.github.context())
}
\arguments{
\item{gist.id}{the gist id}

\item{comment.id}{the comment id}

\item{ctx}{the github context object}
}
\value{
none
}
\description{
Delete gist comment
}

