% Generated by roxygen2 (4.0.2): do not edit by hand
\name{delete.hook}
\alias{delete.hook}
\title{delete a hook.}
\usage{
delete.hook(owner, repo, id, ctx = get.github.context())
}
\arguments{
\item{owner}{the repo owner (user, org, etc)}

\item{repo}{the name of the repo}

\item{id}{the id of the hook}

\item{ctx}{the github context object}
}
\value{
none
}
\description{
delete a hook.
}

