% Generated by roxygen2 (4.0.2): do not edit by hand
\name{delete.member.from.team}
\alias{delete.member.from.team}
\title{remove member from team}
\usage{
delete.member.from.team(id, user, ctx = get.github.context())
}
\arguments{
\item{id}{team id}

\item{user}{the user name}

\item{ctx}{the github context object}
}
\value{
none
}
\description{
remove member from team
}

