% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orgs.R
\name{delete.repository.from.team}
\alias{delete.repository.from.team}
\title{remove repository from team}
\usage{
delete.repository.from.team(id, owner, repo, ctx = get.github.context())
}
\arguments{
\item{id}{team id}

\item{owner}{the repo owner}

\item{repo}{the repo name}

\item{ctx}{the github context object}
}
\value{
none
}
\description{
remove repository from team
}

