% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repositories.R
\name{delete.repository.path}
\alias{delete.repository.path}
\title{Delete a file in a repository}
\usage{
delete.repository.path(owner, repo, path, ..., ctx = get.github.context())
}
\arguments{
\item{owner}{the repo owner (user, org, etc)}

\item{repo}{the name of the repo}

\item{path}{the file path}

\item{...}{extra parameters to be passed. See http://developer.github.com/v3/repos/contents/#update-a-file}

\item{ctx}{the github context object}
}
\value{
the file
}
\description{
Delete a file in a repository
}

