% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_validation.R
\name{.ensure_valid_http_response}
\alias{.ensure_valid_http_response}
\title{Ensure valid HTTP response}
\usage{
.ensure_valid_http_response(resp)
}
\arguments{
\item{resp}{httr response}
}
\value{
NULL. Call for side-effects.
}
\description{
Checks an httr response object and gives appropriate error messages to user.
}
