% Generated by roxygen2 (4.0.2): do not edit by hand
\name{follow.user}
\alias{follow.user}
\title{Start following a given user}
\usage{
follow.user(user, ctx = get.github.context())
}
\arguments{
\item{user}{the given user}

\item{ctx}{the github context object}
}
\value{
none
}
\description{
Start following a given user
}

