% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gists.R
\name{fork.gist}
\alias{fork.gist}
\title{fork a gist}
\usage{
fork.gist(id, ctx = get.github.context())
}
\arguments{
\item{id}{the gist id}

\item{ctx}{the github context object}
}
\value{
the new gist
}
\description{
fork a gist
}

