% Generated by roxygen2 (4.0.2): do not edit by hand
\name{get.gist.comment}
\alias{get.gist.comment}
\title{Get specific comment}
\usage{
get.gist.comment(gist.id, comment.id, ctx = get.github.context())
}
\arguments{
\item{gist.id}{the gist id}

\item{comment.id}{the comment id}

\item{ctx}{the github context object}
}
\value{
the comment
}
\description{
Get specific comment
}

