% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gists.R
\name{get.gist.comments}
\alias{get.gist.comments}
\title{Get comments for a gist}
\usage{
get.gist.comments(id, ctx = get.github.context())
}
\arguments{
\item{id}{the gist id}

\item{ctx}{the github context object}
}
\value{
the list of comments
}
\description{
Get comments for a gist
}

