% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gists.R
\name{get.gist.forks}
\alias{get.gist.forks}
\title{list the forks of a gist}
\usage{
get.gist.forks(id, ctx = get.github.context())
}
\arguments{
\item{id}{the gist id}

\item{ctx}{the github context object}
}
\value{
the list of fork information
}
\description{
list the forks of a gist
}

