\name{get.gist}
\alias{get.gist}
\title{get a specific gist}
\usage{
  get.gist(id, version = NULL, ctx = get.github.context())
}
\arguments{
  \item{id}{the gist id}

  \item{version}{the gist version, optionally}

  \item{ctx}{the github context object}
}
\value{
  the specific gist
}
\description{
  get a specific gist
}

