% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/github.R
\name{get.github.context}
\alias{get.github.context}
\title{returns the most recently created github context, or creates one if none has been so far created}
\usage{
get.github.context()
}
\value{
a github context object
}
\description{
returns the most recently created github context, or creates one if none has been so far created
}

