% Generated by roxygen2 (4.0.2): do not edit by hand
\name{get.milestones}
\alias{get.milestones}
\title{List milestones for a repository}
\usage{
get.milestones(owner, repo, ..., ctx = get.github.context())
}
\arguments{
\item{owner}{the repo owner}

\item{repo}{the repo name}

\item{...}{other parameters. See http://developer.github.com/v3/issues/milestones/#list-milestones-for-a-repository for details.}

\item{ctx}{the github context object}
}
\value{
the milestone list
}
\description{
List milestones for a repository
}

