% Generated by roxygen2 (4.0.2): do not edit by hand
\name{get.my.emails}
\alias{get.my.emails}
\title{Get the list of emails of the current user}
\usage{
get.my.emails(ctx = get.github.context())
}
\arguments{
\item{ctx}{the github context object}
}
\value{
A list of emails
}
\description{
Get the list of emails of the current user
}

