% Generated by roxygen2 (4.0.2): do not edit by hand
\name{get.my.followers}
\alias{get.my.followers}
\title{Get the list of followers for the current user}
\usage{
get.my.followers(ctx = get.github.context())
}
\arguments{
\item{ctx}{the github context object}
}
\value{
the list of followers
}
\description{
Get the list of followers for the current user
}

