% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/users.R
\name{get.my.following}
\alias{get.my.following}
\title{List who is following the current user}
\usage{
get.my.following(ctx = get.github.context())
}
\arguments{
\item{ctx}{the github context object}
}
\value{
the list of followers
}
\description{
List who is following the current user
}

