% Generated by roxygen2 (4.0.2): do not edit by hand
\name{get.my.gists}
\alias{get.my.gists}
\title{list the current user's gists}
\usage{
get.my.gists(ctx = get.github.context())
}
\arguments{
\item{ctx}{the github context object}
}
\value{
the list of gists
}
\description{
list the current user's gists
}

