% Generated by roxygen2 (4.0.2): do not edit by hand
\name{get.my.organizations}
\alias{get.my.organizations}
\title{List all organizations for the current user.}
\usage{
get.my.organizations(ctx = get.github.context())
}
\arguments{
\item{ctx}{the githb context object}
}
\value{
the list of organizations
}
\description{
List all organizations for the current user.
}

