% Generated by roxygen2 (4.0.2): do not edit by hand
\name{get.myself}
\alias{get.myself}
\title{Get information on the current user}
\usage{
get.myself(ctx = get.github.context())
}
\arguments{
\item{ctx}{the github context object}
}
\value{
Information about the user
}
\description{
Get information on the current user
}

