% Generated by roxygen2 (4.0.2): do not edit by hand
\name{get.organization.public.members}
\alias{get.organization.public.members}
\title{list public members of organization}
\usage{
get.organization.public.members(org, ctx = get.github.context())
}
\arguments{
\item{org}{the organization name}

\item{ctx}{the github context object}
}
\value{
the list of public members
}
\description{
list public members of organization
}

