\name{get.organization.teams}
\alias{get.organization.teams}
\title{list teams from an organization}
\usage{
  get.organization.teams(org, ctx = get.github.context())
}
\arguments{
  \item{org}{the organization name}

  \item{ctx}{the github context object}
}
\value{
  the list of organization teams
}
\description{
  list teams from an organization
}

