\name{get.pull.request}
\alias{get.pull.request}
\title{get a specific pull request}
\usage{
  get.pull.request(owner, repo, id,
    ctx = get.github.context())
}
\arguments{
  \item{owner}{the repo owner}

  \item{repo}{the repo name}

  \item{id}{the pull request id}

  \item{ctx}{the github context object}
}
\value{
  the specific pull request
}
\description{
  get a specific pull request
}

