% Generated by roxygen2 (4.0.2): do not edit by hand
\name{get.reference}
\alias{get.reference}
\title{get a git reference}
\usage{
get.reference(owner, repo, ref, ctx = get.github.context())
}
\arguments{
\item{owner}{the repo owner}

\item{repo}{the name of the repo}

\item{ref}{the git reference}

\item{ctx}{the github context object}
}
\value{
the reference
}
\description{
get a git reference
}

