% Generated by roxygen2 (4.0.2): do not edit by hand
\name{get.repository.branches}
\alias{get.repository.branches}
\title{get list of repository branches}
\usage{
get.repository.branches(owner, repo, ctx = get.github.context())
}
\arguments{
\item{owner}{the repo owner (user, org, etc)}

\item{repo}{the name of the repo}

\item{ctx}{the github context object}
}
\value{
list of branches
}
\description{
get list of repository branches
}

