\name{get.repository.contributors}
\alias{get.repository.contributors}
\title{get list of repository contributors}
\usage{
  get.repository.contributors(owner, repo,
    ctx = get.github.context())
}
\arguments{
  \item{owner}{the repo owner (user, org, etc)}

  \item{repo}{the name of the repo}

  \item{ctx}{the github context object}
}
\value{
  list of repo contributors
}
\description{
  get list of repository contributors
}

