% Generated by roxygen2 (4.0.2): do not edit by hand
\name{get.repository.contributors.stats}
\alias{get.repository.contributors.stats}
\title{get stats on repository contributors}
\usage{
get.repository.contributors.stats(owner, repo, ctx = get.github.context())
}
\arguments{
\item{owner}{the repo owner (user, org, etc)}

\item{repo}{the name of the repo}

\item{ctx}{the github context object}
}
\value{
list of repo contributors stats
}
\description{
get stats on repository contributors
}

