% Generated by roxygen2 (4.0.2): do not edit by hand
\name{get.repository.downloads}
\alias{get.repository.downloads}
\title{list downloads for a repository}
\usage{
get.repository.downloads(owner, repo, ctx = get.github.context())
}
\arguments{
\item{owner}{the repo owner (user, org, etc)}

\item{repo}{the name of the repo}

\item{ctx}{the github context object}
}
\value{
the list of downloads
}
\description{
list downloads for a repository
}

