% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repositories.R
\name{get.repository.key}
\alias{get.repository.key}
\title{get a specific repository key}
\usage{
get.repository.key(owner, repo, id, ctx = get.github.context())
}
\arguments{
\item{owner}{the repo owner (user, org, etc)}

\item{repo}{the name of the repo}

\item{id}{the key id}

\item{ctx}{the github context object}
}
\value{
the key
}
\description{
get a specific repository key
}

