\name{get.repository.label}
\alias{get.repository.label}
\title{get a specific label for a repository}
\usage{
  get.repository.label(owner, repo, name,
    ctx = get.github.context())
}
\arguments{
  \item{owner}{the repo owner}

  \item{repo}{the repo name}

  \item{name}{the label name}

  \item{ctx}{the github context object}
}
\value{
  The given label
}
\description{
  get a specific label for a repository
}

