% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/issues.R
\name{get.repository.labels}
\alias{get.repository.labels}
\title{List all labels for a repository}
\usage{
get.repository.labels(owner, repo, ctx = get.github.context())
}
\arguments{
\item{owner}{the repo owner}

\item{repo}{the repo name}

\item{ctx}{the github context object}
}
\value{
The list of labels
}
\description{
List all labels for a repository
}

