% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repositories.R
\name{get.repository}
\alias{get.repository}
\title{get repository}
\usage{
get.repository(owner, repo, ctx = get.github.context())
}
\arguments{
\item{owner}{the repo owner (user, org, etc)}

\item{repo}{the name of the repo}

\item{ctx}{the github context object}
}
\value{
the repository
}
\description{
get repository
}

