% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repositories.R
\name{get.repository.readme}
\alias{get.repository.readme}
\title{Get the README for a repo}
\usage{
get.repository.readme(owner, repo, ..., ctx = get.github.context())
}
\arguments{
\item{owner}{the repo owner (user, org, etc)}

\item{repo}{the name of the repo}

\item{...}{extra parameters to be passed. See \url{http://developer.github.com/v3/repos/contents/#get-the-readme} for details.}

\item{ctx}{the github context object}
}
\value{
the readme
}
\description{
Get the README for a repo
}

