% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repositories.R
\name{get.repository.status}
\alias{get.repository.status}
\title{get combined status for a ref in a repo}
\usage{
get.repository.status(owner, repo, ref, ctx = get.github.context())
}
\arguments{
\item{owner}{the repo owner (user, org, etc)}

\item{repo}{the name of the repo}

\item{ref}{Ref to list the statuses from. It can be a SHA, a branch name, or a tag name.}

\item{ctx}{the github context object}
}
\value{
the combined status
}
\description{
get combined status for a ref in a repo
}

