% Generated by roxygen2 (4.0.2): do not edit by hand
\name{get.team}
\alias{get.team}
\title{get team information}
\usage{
get.team(id, ctx = get.github.context())
}
\arguments{
\item{id}{the id of the team}

\item{ctx}{the github context object}
}
\value{
team information
}
\description{
get team information
}

