% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activity.R
\name{get.thread.notifications}
\alias{get.thread.notifications}
\title{get single thread notifications}
\usage{
get.thread.notifications(id, ctx = get.github.context())
}
\arguments{
\item{id}{the thread id}

\item{ctx}{the github context object}
}
\value{
the thread
}
\description{
get single thread notifications
}

