\name{get.user.gists}
\alias{get.user.gists}
\title{list the gists of a given user}
\usage{
  get.user.gists(user, ctx = get.github.context())
}
\arguments{
  \item{user}{the given user}

  \item{ctx}{the github context object}
}
\value{
  the list of gists
}
\description{
  list the gists of a given user
}

