% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/users.R
\name{get.user.keys}
\alias{get.user.keys}
\title{Get public keys for a given user}
\usage{
get.user.keys(user, ctx = get.github.context())
}
\arguments{
\item{user}{the given user}

\item{ctx}{the github context object}
}
\value{
list of keys
}
\description{
Get public keys for a given user
}

