% Generated by roxygen2 (4.0.2): do not edit by hand
\name{get.user.public.received.events}
\alias{get.user.public.received.events}
\title{list public events that a user has received}
\usage{
get.user.public.received.events(user, ..., ctx = get.github.context())
}
\arguments{
\item{user}{the user}

\item{...}{extra parameters. See http://developer.github.com/v3/activity/events/ for details}

\item{ctx}{the github context object}
}
\value{
the list of public events the user has received
}
\description{
list public events that a user has received
}

