% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activity.R
\name{get.watchers}
\alias{get.watchers}
\title{list repo watchers}
\usage{
get.watchers(owner, repo, ctx = get.github.context())
}
\arguments{
\item{owner}{repo owner}

\item{repo}{name of repo}

\item{ctx}{the github context object}
}
\value{
list of repository watchers
}
\description{
list repo watchers
}

