% Generated by roxygen2 (4.0.2): do not edit by hand
\name{is.member.in.team}
\alias{is.member.in.team}
\title{test if user is a member of team}
\usage{
is.member.in.team(id, user, ctx = get.github.context())
}
\arguments{
\item{id}{team id}

\item{user}{the user name}

\item{ctx}{the github context object}
}
\value{
TRUE if user is member of team
}
\description{
test if user is a member of team
}

