% Generated by roxygen2 (4.0.2): do not edit by hand
\name{is.member.of.organization}
\alias{is.member.of.organization}
\title{test if user is a member of an organization}
\usage{
is.member.of.organization(org, user, ctx = get.github.context())
}
\arguments{
\item{org}{the organization name}

\item{user}{the user name}

\item{ctx}{the github context object}
}
\value{
TRUE if user is a member of organization
}
\description{
test if user is a member of an organization
}

