% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activity.R
\name{mark.my.repository.notifications}
\alias{mark.my.repository.notifications}
\title{mark my notifications as read for a given repo}
\usage{
mark.my.repository.notifications(owner, repo, ..., ctx = get.github.context())
}
\arguments{
\item{owner}{the repo owner}

\item{repo}{the name of the repo}

\item{...}{extra parameters. See \url{http://developer.github.com/v3/activity/notifications/#mark-notifications-as-read-in-a-repository} for}

\item{ctx}{the github context object}
}
\value{
none
}
\description{
mark my notifications as read for a given repo
}

