% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activity.R
\name{mark.thread.notifications}
\alias{mark.thread.notifications}
\title{mark a single thread as read.}
\usage{
mark.thread.notifications(id, ctx = get.github.context())
}
\arguments{
\item{id}{the thread id}

\item{ctx}{the github context object}
}
\value{
none
}
\description{
mark a single thread as read.
}

