% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repositories.R
\name{modify.hook}
\alias{modify.hook}
\title{modify an exisitng hook for repo}
\usage{
modify.hook(owner, repo, id, content, ctx = get.github.context())
}
\arguments{
\item{owner}{the repo owner (user, org, etc)}

\item{repo}{the name of the repo}

\item{id}{the id of the hook}

\item{content}{the JSON object describing the hook. See \url{http://developer.github.com/v3/repos/hooks/#edit-a-hook} for details.}

\item{ctx}{the github context object}
}
\value{
the specific hook
}
\description{
modify an exisitng hook for repo
}

