% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/users.R
\name{modify.key}
\alias{modify.key}
\title{Update a public key with some id}
\usage{
modify.key(id, content, ctx = get.github.context())
}
\arguments{
\item{id}{the public key id}

\item{content}{the public key}

\item{ctx}{the github context object}
}
\value{
the updated public key
}
\description{
Update a public key with some id
}

