% Generated by roxygen2 (4.0.2): do not edit by hand
\name{perform.merge.pull.request}
\alias{perform.merge.pull.request}
\title{merge a pull request}
\usage{
perform.merge.pull.request(owner, repo, id, ..., ctx = get.github.context())
}
\arguments{
\item{owner}{the repo owner}

\item{repo}{the repo name}

\item{id}{the pull request id}

\item{...}{extra parameters to control the merge. See http://developer.github.com/v3/pulls/#merge-a-pull-request-merge-buttontrade for details.}

\item{ctx}{the github context object}
}
\value{
none
}
\description{
merge a pull request
}

