% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repositories.R
\name{perform.repository.merge}
\alias{perform.repository.merge}
\title{perform a merge.}
\usage{
perform.repository.merge(owner, repo, content, ctx = get.github.context())
}
\arguments{
\item{owner}{the repo owner (user, org, etc)}

\item{repo}{the name of the repo}

\item{content}{the JSON object describing the merge. See \url{http://developer.github.com/v3/repos/merging/#perform-a-merge} for details.}

\item{ctx}{the github context object}
}
\value{
the resulting merge commit
}
\description{
perform a merge.
}

