% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/issues.R
\name{replace.issue.labels}
\alias{replace.issue.labels}
\title{Replace all issue labels}
\usage{
replace.issue.labels(owner, repo, issue.number, content,
  ctx = get.github.context())
}
\arguments{
\item{owner}{the repo owner}

\item{repo}{the repo name}

\item{issue.number}{the number of the issue}

\item{content}{the new list of labels}

\item{ctx}{the github context object}
}
\value{
The new list of labels
}
\description{
Replace all issue labels
}

