% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Repository.R
\name{Repository}
\alias{Repository}
\title{R6 Class for GitHub repositories}
\description{
Repository class representing GitHub API response for that repo.
}
\details{
Contains three fields \strong{owner}, \strong{name} of the
repository, and \strong{info} field, which represents response from GitHub API, in
form of a list. This class can later be used to change repo parameters (such
as: description, collaborators etc.)
}
\examples{
repo <- Repository$new(owner = "tidyverse", repo = "dplyr")
repo <- Repository$new("tidyverse", "dplyr")
\dontrun{
repo <- Repository$new(owner = "owner", repo = "repo", token = "token")
}

}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{owner}}{string Owner of the GitHub repo.}

\item{\code{repo}}{string Name of the Github repo.}

\item{\code{info}}{list Info field which will contain response with all
additional information about repo.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Repository$new()}}
\item \href{#method-print}{\code{Repository$print()}}
\item \href{#method-update_description}{\code{Repository$update_description()}}
\item \href{#method-clone}{\code{Repository$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a new repository object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Repository$new(owner, repo, info = NULL, token = NA)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{owner}}{string Owner of the repo.}

\item{\code{repo}}{string Name of the repo.}

\item{\code{info}}{list List of info fields.}

\item{\code{token}}{string Token, not a mandatory field.
needed for private repos.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new `Repository` object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\if{latex}{\out{\hypertarget{method-print}{}}}
\subsection{Method \code{print()}}{
Change object printing
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Repository$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-update_description"></a>}}
\if{latex}{\out{\hypertarget{method-update_description}{}}}
\subsection{Method \code{update_description()}}{
Update repository description
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Repository$update_description(description = NA)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{description}}{string New description.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new `Repository` object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Repository$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
