% Generated by roxygen2 (4.0.2): do not edit by hand
\name{search.code}
\alias{search.code}
\title{Search the Github API.}
\usage{
search.code(q, ..., ctx = get.github.context())
}
\arguments{
\item{q}{Query string}

\item{...}{Further arguments passed on to Github API call}

\item{ctx}{Authentication object}
}
\description{
Search the Github API.
}
\examples{
\dontrun{
search.code("octokit in:file extension:gemspec -repo:octokit/octokit.rb", sort="indexed")
}
}

