% Generated by roxygen2 (4.0.2): do not edit by hand
\name{search.issues}
\alias{search.issues}
\title{Search the Github API.}
\usage{
search.issues(q, ..., ctx = get.github.context())
}
\arguments{
\item{q}{Query string}

\item{...}{Further arguments passed on to Github API call}

\item{ctx}{Authentication object}
}
\description{
Search the Github API.
}
\examples{
\dontrun{
search.issues("windows label:bug language:python state:open", sort="created", order="asc")
}
}

