% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{search.repositories}
\alias{search.repositories}
\title{Search Github repositories.}
\usage{
search.repositories(q, ..., ctx = get.github.context())
}
\arguments{
\item{q}{Query string}

\item{...}{Further arguments passed on to Github API call}

\item{ctx}{Authentication object}
}
\description{
Search Github repositories.

Search the Github API.
}
\examples{
\dontrun{
search.repositories("tetris language:assembly")
}
}

