% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gists.R
\name{star.gist}
\alias{star.gist}
\title{star a gist}
\usage{
star.gist(id, ctx = get.github.context())
}
\arguments{
\item{id}{the gist id}

\item{ctx}{the github context object}
}
\value{
none
}
\description{
star a gist
}

