% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/users.R
\name{unfollow.user}
\alias{unfollow.user}
\title{Stop following a given user}
\usage{
unfollow.user(user, ctx = get.github.context())
}
\arguments{
\item{user}{the given user}

\item{ctx}{the github context object}
}
\value{
none
}
\description{
Stop following a given user
}

