% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repositories.R
\name{update.repository.path}
\alias{update.repository.path}
\title{Update the contents of a file}
\usage{
\method{update}{repository.path}(owner, repo, path, ...,
  ctx = get.github.context())
}
\arguments{
\item{owner}{the repo owner (user, org, etc)}

\item{repo}{the name of the repo}

\item{path}{the file path}

\item{...}{extra parameters to be passed. See \url{http://developer.github.com/v3/repos/contents/#update-a-file}}

\item{ctx}{the github context object}
}
\value{
the file
}
\description{
Update the contents of a file
}

