% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/User.R
\name{User}
\alias{User}
\title{R6 Class for GitHub users}
\description{
User class representing GitHub API response for that user.
}
\details{
Contains two fields \strong{username} and \strong{info} field, which
represents response from GitHub API, in form of a list. This class can later
be used to change repo parameters (such as: description, collaborators etc.)
}
\examples{
user <- User$new(username = "tidyverse")
user <- User$new("tidyverse")
\dontrun{
user <- User$new(username = "username")
}

}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{username}}{string Username of the GitHub user account.}

\item{\code{info}}{list Info field which will contain response with all
additional information about user.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{User$new()}}
\item \href{#method-print}{\code{User$print()}}
\item \href{#method-clone}{\code{User$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a new user object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{User$new(username = NA)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{username}}{string Username of the GitHub account.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new `User` object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\if{latex}{\out{\hypertarget{method-print}{}}}
\subsection{Method \code{print()}}{
Change object printing
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{User$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{User$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
