% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{create.blob}
\alias{create.blob}
\title{create a blob}
\usage{
create.blob(owner, repo, content, ctx = get.github.context())
}
\arguments{
\item{owner}{the repo owner}

\item{repo}{the name of the repo}

\item{content}{the JSON object describing the blob. See \url{http://developer.github.com/v3/git/blobs/#create-a-blob} for details.}

\item{ctx}{the github context object}
}
\value{
the blob
}
\description{
create a blob
}

