% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gists.R
\name{create.gist.comment}
\alias{create.gist.comment}
\title{Create gist comment}
\usage{
create.gist.comment(gist.id, content, ctx = get.github.context())
}
\arguments{
\item{gist.id}{the gist id}

\item{content}{the JSON object describing the comment. see \url{http://developer.github.com/v3/gists/comments/#create-a-comment} for details}

\item{ctx}{the github context object}
}
\value{
the comment
}
\description{
Create gist comment
}

