% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/users.R
\name{delete.key}
\alias{delete.key}
\title{Delete a public key with some id}
\usage{
delete.key(id, ctx = get.github.context())
}
\arguments{
\item{id}{the public key id}

\item{ctx}{the github context object}
}
\value{
None
}
\description{
Delete a public key with some id
}

