\name{delete.reference}
\alias{delete.reference}
\title{delete a reference}
\usage{
  delete.reference(owner, repo, ref,
    ctx = get.github.context())
}
\arguments{
  \item{owner}{the repo owner}

  \item{repo}{the name of the repo}

  \item{ref}{the reference}

  \item{ctx}{the github context object}
}
\value{
  none
}
\description{
  delete a reference
}

