\name{delete.repository}
\alias{delete.repository}
\title{delete repository}
\usage{
  delete.repository(owner, repo,
    ctx = get.github.context())
}
\arguments{
  \item{owner}{the repo owner (user, org, etc)}

  \item{repo}{the name of the repo}

  \item{ctx}{the github context object}
}
\value{
  nothing
}
\description{
  delete repository
}

