% Generated by roxygen2 (4.0.2): do not edit by hand
\name{delete.team}
\alias{delete.team}
\title{delete a team in an organization}
\usage{
delete.team(id, ctx = get.github.context())
}
\arguments{
\item{id}{team id}

\item{ctx}{the github context object}
}
\value{
none
}
\description{
delete a team in an organization
}

