\name{get.assignees}
\alias{get.assignees}
\title{Get list of assignees for a repo}
\usage{
  get.assignees(owner, repo, ctx = get.github.context())
}
\arguments{
  \item{owner}{the repo owner}

  \item{repo}{the repo name}

  \item{ctx}{the github context object}
}
\value{
  the list of assignees
}
\description{
  Get list of assignees for a repo
}

