\name{get.blob}
\alias{get.blob}
\title{get a blob}
\usage{
  get.blob(owner, repo, sha, ctx = get.github.context())
}
\arguments{
  \item{owner}{the repo owner}

  \item{repo}{the name of the repo}

  \item{sha}{the SHA of the blob}

  \item{ctx}{the github context object}
}
\value{
  the blob
}
\description{
  get a blob
}

