% Generated by roxygen2 (4.0.2): do not edit by hand
\name{get.my.keys}
\alias{get.my.keys}
\title{Get public keys for the current user}
\usage{
get.my.keys(ctx = get.github.context())
}
\arguments{
\item{ctx}{the github context object}
}
\value{
list of keys
}
\description{
Get public keys for the current user
}

