% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/users.R
\name{get.myself}
\alias{get.myself}
\title{Get information on the current user}
\usage{
get.myself(ctx = get.github.context())
}
\arguments{
\item{ctx}{the github context object}
}
\value{
Information about the user
}
\description{
Get information on the current user
}

