\name{get.organization}
\alias{get.organization}
\title{Get an organization}
\usage{
  get.organization(org, ctx = get.github.context())
}
\arguments{
  \item{org}{the organization name}

  \item{ctx}{the githb context object}
}
\value{
  the organization information
}
\description{
  Get an organization
}

