% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/pulls.R
\name{get.pull.request.commits}
\alias{get.pull.request.commits}
\title{list commits for a pull request}
\usage{
get.pull.request.commits(owner, repo, id, ..., ctx = get.github.context())
}
\arguments{
\item{owner}{the repo owner}

\item{repo}{the repo name}

\item{id}{the pull request id}

\item{...}{extra parameters. See
https://developer.github.com/v3/pulls/#list-commits-on-a-pull-request for details}

\item{ctx}{the github context object}
}
\value{
the list of pull request commits
}
\description{
list commits for a pull request
}

