\name{get.repositories.watched.by.user}
\alias{get.repositories.watched.by.user}
\title{list repos watched by users}
\usage{
  get.repositories.watched.by.user(user,
    ctx = get.github.context())
}
\arguments{
  \item{user}{the user}

  \item{ctx}{the github context object}
}
\value{
  list of repositories
}
\description{
  list repos watched by users
}

