% Generated by roxygen2 (4.0.2): do not edit by hand
\name{get.repository.keys}
\alias{get.repository.keys}
\title{list all repository keys}
\usage{
get.repository.keys(owner, repo, ctx = get.github.context())
}
\arguments{
\item{owner}{the repo owner (user, org, etc)}

\item{repo}{the name of the repo}

\item{ctx}{the github context object}
}
\value{
the key list
}
\description{
list all repository keys
}

