% Generated by roxygen2 (4.0.2): do not edit by hand
\name{get.repository}
\alias{get.repository}
\title{get repository}
\usage{
get.repository(owner, repo, ctx = get.github.context())
}
\arguments{
\item{owner}{the repo owner (user, org, etc)}

\item{repo}{the name of the repo}

\item{ctx}{the github context object}
}
\value{
the repository
}
\description{
get repository
}

