% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orgs.R
\name{get.team.members}
\alias{get.team.members}
\title{list team members}
\usage{
get.team.members(id, ctx = get.github.context())
}
\arguments{
\item{id}{team id}

\item{ctx}{the github context object}
}
\value{
team member list
}
\description{
list team members
}

