\name{get.thread.notifications}
\alias{get.thread.notifications}
\title{get single thread notifications}
\usage{
  get.thread.notifications(id, ctx = get.github.context())
}
\arguments{
  \item{id}{the thread id}

  \item{ctx}{the github context object}
}
\value{
  the thread
}
\description{
  get single thread notifications
}

