\name{get.thread.notifications.subscription}
\alias{get.thread.notifications.subscription}
\title{checks to see if the current user is subscribed to a thread.}
\usage{
  get.thread.notifications.subscription(id,
    ctx = get.github.context())
}
\arguments{
  \item{id}{the thread id}

  \item{ctx}{the github context object}
}
\value{
  the response object
}
\description{
  checks to see if the current user is subscribed to a
  thread.
}

