% Generated by roxygen2 (4.0.2): do not edit by hand
\name{get.user.following}
\alias{get.user.following}
\title{List who a user is following}
\usage{
get.user.following(user, ctx = get.github.context())
}
\arguments{
\item{user}{the user}

\item{ctx}{the github context object}
}
\value{
the list of followers
}
\description{
List who a user is following
}

