\name{get.user.keys}
\alias{get.user.keys}
\title{Get public keys for a given user}
\usage{
  get.user.keys(user, ctx = get.github.context())
}
\arguments{
  \item{user}{the given user}

  \item{ctx}{the github context object}
}
\value{
  list of keys
}
\description{
  Get public keys for a given user
}

