% Generated by roxygen2 (4.0.2): do not edit by hand
\name{get.user}
\alias{get.user}
\title{Get information on some user}
\usage{
get.user(user, ctx = get.github.context())
}
\arguments{
\item{user}{the user}

\item{ctx}{the github context object}
}
\value{
Information about the user
}
\description{
Get information on some user
}

