% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{package}
\name{github}
\alias{github}
\alias{github-package}
\title{github-package: use the Github API from R}
\description{
This package wraps the Github web service API so you can make R
  calls against the Github API (to get information about repositories, or
  even to create new content)
}
\examples{
\dontrun{get.user.repositories("cscheid")}
}
\author{
Carlos Scheidegger
}
\seealso{
\code{RJSON}
}
\keyword{github-package}
\keyword{package}

