% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/github.R
\name{interactive.login}
\alias{interactive.login}
\title{Obtain a github context interactively}
\usage{
interactive.login(client_id = NULL, client_secret = NULL, scopes = NULL,
  base_url = "https://github.com", api_url = "https://api.github.com",
  max_etags = 10000, verbose = FALSE)
}
\arguments{
\item{client_id}{the github client ID}

\item{client_secret}{the github client secret}

\item{scopes}{the OAuth scopes you want to request}

\item{base_url}{the base URL for the github webpage. Change this in GitHub
Enterprise deployments to your base G.E. URL}

\item{api_url}{the base URL for the github API. Change this in GitHub
Enterprise deployments to your base G.E. API URL}

\item{max_etags}{the maximum number of entries to cache in the context}

\item{verbose}{logical, passed to \code{create.github.context} and,
ultimately, to httr configuration}
}
\value{
a github context object that is used in every github API call issued
  by this library.
}
\description{
interactive.login opens a web browser, asks for your username+password,
performs the OAuth dance, retrieves the token, and uses it to create a github
context.
}
\details{
If you use rgithub interactively, then you will get a default
client ID and client secret that you can use. Please don't abuse that,
or the feature might need to be removed.

Refer to http://developer.github.com/guides/basics-of-authentication/
}

