\name{mark.my.notifications}
\alias{mark.my.notifications}
\title{mark my notifications as read}
\usage{
  mark.my.notifications(..., ctx = get.github.context())
}
\arguments{
  \item{...}{extra parameters. See
  http://developer.github.com/v3/activity/notifications/#mark-notifications-as-read}

  \item{ctx}{the github context object}
}
\value{
  None
}
\description{
  mark my notifications as read
}

