% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orgs.R
\name{modify.team}
\alias{modify.team}
\title{edit a team in an organization}
\usage{
modify.team(id, content, ctx = get.github.context())
}
\arguments{
\item{id}{team id}

\item{content}{the JSON object describing the team. See \url{http://developer.github.com/v3/orgs/teams/#create-team} for details.}

\item{ctx}{the github context object}
}
\value{
team information
}
\description{
edit a team in an organization
}

