% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orgs.R
\name{publicize.member.of.organization}
\alias{publicize.member.of.organization}
\title{publicize user membership in an organization}
\usage{
publicize.member.of.organization(org, user, ctx = get.github.context())
}
\arguments{
\item{org}{the organization name}

\item{user}{the user name}

\item{ctx}{the github context object}
}
\value{
None
}
\description{
publicize user membership in an organization
}

