\name{rGitHub-package}
\alias{rGitHub-package}
\alias{rGitHub}
\docType{package}
\title{
\packageTitle{rGitHub}
}
\description{
\packageDescription{rGitHub}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{rGitHub}
\packageIndices{rGitHub}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{rGitHub}

Maintainer: \packageMaintainer{rGitHub}
}
\references{
~~ Literature or other references for background information ~~
}

\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
