% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rGithub-package.R
\docType{package}
\name{rGithub-package}
\alias{rGithub}
\alias{rGithub-package}
\title{rGithub: R Package for Github REST API}
\description{
Small client library to access Github's REST API.
}
\author{
\strong{Maintainer}: Vladimir Obucina \email{vlada_3292@hotmail.com}

}
\keyword{internal}
