% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{search.issues}
\alias{search.issues}
\title{Search Github issues.}
\usage{
search.issues(q, ..., ctx = get.github.context())
}
\arguments{
\item{q}{Query string}

\item{...}{Further arguments passed on to Github API call}

\item{ctx}{Authentication object}
}
\description{
Search Github issues.

Search the Github API.
}
\examples{
\dontrun{
search.issues("windows label:bug language:python state:open", sort="created", order="asc")
}
}

