% Generated by roxygen2 (4.0.2): do not edit by hand
\name{search.repositories}
\alias{search.repositories}
\title{Search the Github API.}
\usage{
search.repositories(q, ..., ctx = get.github.context())
}
\arguments{
\item{q}{Query string}

\item{...}{Further arguments passed on to Github API call}

\item{ctx}{Authentication object}
}
\description{
Search the Github API.
}
\examples{
\dontrun{
search.repositories("tetris language:assembly")
}
}

