% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{search.users}
\alias{search.users}
\title{Search Github users.}
\usage{
search.users(q, ..., ctx = get.github.context())
}
\arguments{
\item{q}{Query string}

\item{...}{Further arguments passed on to Github API call}

\item{ctx}{Authentication object}
}
\description{
Search Github users.

Search the Github API.
}
\examples{
\dontrun{
search.users("tom repos:>42 followers:>1000")
}
}

