% Generated by roxygen2 (4.0.2): do not edit by hand
\name{search.users}
\alias{search.users}
\title{Search the Github API.}
\usage{
search.users(q, ..., ctx = get.github.context())
}
\arguments{
\item{q}{Query string}

\item{...}{Further arguments passed on to Github API call}

\item{ctx}{Authentication object}
}
\description{
Search the Github API.
}
\examples{
\dontrun{
search.users("tom repos:>42 followers:>1000")
}
}

