\name{star.repository}
\alias{star.repository}
\title{star a repository}
\usage{
  star.repository(owner, repo, ctx = get.github.context())
}
\arguments{
  \item{owner}{repo owner}

  \item{repo}{name of repo}

  \item{ctx}{the github context object}
}
\value{
  none
}
\description{
  star a repository
}

