\name{unset.repository.subscription}
\alias{unset.repository.subscription}
\title{clear repository subscription info}
\usage{
  unset.repository.subscription(owner, repo,
    ctx = get.github.context())
}
\arguments{
  \item{owner}{repo owner}

  \item{repo}{name of repo}

  \item{ctx}{the github context object}
}
\value{
  subscription info
}
\description{
  clear repository subscription info
}

